Changed the typedef in types.h back to short.  Changed line 21 of syscalls.h so the offset is 62 rather than 64, like in Bander's code.

+++ include/utils/types.h
@@ -4,7 +4,7 @@
 
 #ifndef HIGHPERFORMANCELINUXSENSORS_TYPES_H
 #define HIGHPERFORMANCELINUXSENSORS_TYPES_H
-typedef int pid_t;
-typedef unsigned long int syscall_t;
+typedef short pid_t;
+typedef short syscall_t;
 //typedef int syscall_t;
 #endif //HIGHPERFORMANCELINUXSENSORS_TYPES_H
